+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.19] - 2018-10-18
+### Added
+- SMSG_FRIEND_LIST
+- SMSG_IGNORE_LIST
+- FriendResults (Enum)
+- SubFriends Character Database.
+- SubIgnoreds Character Database.
+- MOTD Support.
+- OnFriendList.
+- OnMessageChat *Partial.
+- OnFriendAdd *Partial.
+
 ## [0.0.18] - 2018-10-18
 ### Added
 - CharacterHelper.GetRaceModel